修改通信协议

+++ light/protocol/m_protocol.go
@@ -2,11 +2,11 @@ package protocol
 
 // 消息类型
 const (
-	C_TYPE_HANDSHAKE     uint8 = iota + 1 // 握手消息
-	C_TYPE_HANDSHAKE_ACK                  // 客户端->服务器握手ACK
-	C_TYPE_HEARTBEAT                      // 心跳消息
-	C_TYPE_DATA                           // 数据
-	C_TYPE_KICK                           // 服务器主动断开
+	C_TYPE_HANDSHAKE     uint16 = iota + 1 // 握手消息
+	C_TYPE_HANDSHAKE_ACK                   // 客户端->服务器握手ACK
+	C_TYPE_HEARTBEAT                       // 心跳消息
+	C_TYPE_DATA                            // 数据
+	C_TYPE_KICK                            // 服务器主动断开
 )
 
 // 主协议常量
